move pages' components to components dir
