+++ src/components/pages/bookshelf/drawer.tsx
@@ -1,9 +1,9 @@
 import type * as Prisma from "@prisma/client";
-import { drawerWidth } from ".";
+import { drawerWidth } from "../../../pages/bookshelf";
 import { makeStyles } from "y/utils/makesStyles";
 import { useState } from "react";
 import { useRouter } from "next/router";
-import { apiLogout } from "../clientApi";
+import { apiLogout } from "../../../pages/clientApi";
 import Divider from "@mui/material/Divider";
 import List from "@mui/material/List";
 import ListItem from "@mui/material/ListItem";
